Add files via upload
Two images of red blood cells taken from <https://journals.plos.org/plosone/article/figure?id=10.1371/journal.pone.0215447.g001>
The python script converts the jpeg images to an array, performs a simple pixel averaging and then a simple smoothing of a row of pixels. It plots a graph of the row array and then a smoothed version of that. From there you can calculate the size of the red blood cell.